02/requirements.txt
requests
fastapi[standard]
python-dotenv
